+++ Cargo.toml
@@ -7,6 +7,7 @@ edition = "2018"
 [dependencies]
 winapi = { version = "0.3", features = [ "impl-default", "errhandlingapi", "fileapi", "minwindef", "wintrust" ] }
 cellophane = { git = "https://github.com/rustysec/cellophane-rs" }
+serde = { version = "1", features = [ "derive" ], optional = true }
 thiserror = "1"
 widestring = "0.4"
 
@@ -16,3 +17,4 @@ bindgen = { version = "0.58", optional = true }
 
 [features]
 with-bindgen = [ "bindgen" ]
+serialize = [ "serde" ]
